requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sentence-transformers==2.6.1
chromadb==0.5.3

discord.py==2.3.2
python-dotenv==1.0.1
requests==2.31.0

# Optional but recommended for local model performance
numpy==1.26.4
scipy==1.12.0
pandas==2.2.0